CloseButton 멤버 변수 이동
OptionPopUpWidget -> PopUpWidgetBase
