+++ Source/MultiplayerTetris/Private/OptionPopUpWidget.cpp
@@ -1,20 +1,9 @@
 // Copyright Ryu KeunBeom. All Rights Reserved.
 
 #include "OptionPopUpWidget.h"
-#include "ButtonBase.h"
 
 void UOptionPopUpWidget::NativeConstruct()
 {
 	Super::NativeConstruct();
 
-	if (!CloseButton->OnClicked.IsBound())
-	{
-		CloseButton->OnClicked.AddDynamic(this, &UOptionPopUpWidget::OnCloseClicked);
-	}
-}
-
-void UOptionPopUpWidget::OnCloseClicked()
-{
-	// Close this Widget
-	RemoveFromParent();
 }
