+++ Source/MultiplayerTetris/Public/PopUpWidgetBase.h
@@ -6,6 +6,8 @@
 #include "Blueprint/UserWidget.h"
 #include "PopUpWidgetBase.generated.h"
 
+class UButtonBase;
+
 /**
  * 
  */
@@ -13,5 +15,15 @@ UCLASS()
 class MULTIPLAYERTETRIS_API UPopUpWidgetBase : public UUserWidget
 {
 	GENERATED_BODY()
+
+protected:
+	virtual void NativeConstruct() override;
 	
+private:
+	UFUNCTION()
+	void OnCloseClicked();
+
+private:
+	UPROPERTY(meta = (BindWidget))
+	TObjectPtr<UButtonBase> CloseButton;
 };
